refactor(ui): extract EventBus to module, replace hooks.renderTodos with EventBus dispatch (task 154)
- Extract EventBus singleton from inline app.js IIFE into client/modules/eventBus.js
- Replace all 34 hooks.renderTodos?.() calls across 7 domain modules with
  EventBus.dispatch("todos:changed", { reason }) — business logic now signals
  state changes rather than commanding render sweeps
- Remove now-unused hooks.renderTodos assignment from app.js wireHooks
- Each dispatch carries a descriptive reason string for auditability

Stats: 34 calls replaced, 9 files changed, 1 new module
