+++ client/modules/overlayManager.js
@@ -6,6 +6,7 @@
 // =============================================================================
 
 import { state, hooks } from "./store.js";
+import { EventBus } from "./eventBus.js";
 
 // ---------------------------------------------------------------------------
 // Confirm dialog
@@ -201,7 +202,7 @@ export async function saveEditedTodo() {
 
   try {
     await hooks.applyTodoPatch(state.editingTodoId, payload);
-    hooks.renderTodos?.();
+    EventBus.dispatch("todos:changed", { reason: "todo-updated" });
     hooks.syncTodoDrawerStateWithRender?.();
     closeEditTodoModal();
     showMessage("todosMessage", "Task updated", "success");
