+++ client/modules/todosService.js
@@ -3,6 +3,7 @@
 // Imports state from store.js. Cross-module calls go through hooks.
 // =============================================================================
 import { state, hooks, createInitialHomeTopFocusState } from "./store.js";
+import { EventBus } from "./eventBus.js";
 
 // ---------------------------------------------------------------------------
 // Helpers — apiCall, API_URL, normalizeProjectPath, parseApiBody are injected
@@ -161,7 +162,7 @@ async function loadVisibleTodos({ force = false } = {}) {
   visibleTodosState.requestSeq = requestSeq;
   visibleTodosState.loading = true;
   visibleTodosState.pendingQueryKey = queryKey;
-  hooks.renderTodos?.();
+  EventBus.dispatch("todos:changed", { reason: "todos-loading" });
 
   try {
     const todosUrl = hooks.buildUrl(`${hooks.API_URL}/todos`, queryParams);
@@ -173,7 +174,7 @@ async function loadVisibleTodos({ force = false } = {}) {
       visibleTodosState.items = await response.json();
       visibleTodosState.queryKey = queryKey;
       visibleTodosState.loading = false;
-      hooks.renderTodos?.();
+      EventBus.dispatch("todos:changed", { reason: "todos-loaded" });
       return true;
     }
   } catch (error) {
@@ -182,7 +183,7 @@ async function loadVisibleTodos({ force = false } = {}) {
 
   if (requestSeq === visibleTodosState.requestSeq) {
     clearVisibleTodosState();
-    hooks.renderTodos?.();
+    EventBus.dispatch("todos:changed", { reason: "todos-load-error" });
   }
   return false;
 }
@@ -198,7 +199,7 @@ async function refreshVisibleTodosIfNeeded() {
 async function loadTodos() {
   state.todosLoadState = "loading";
   state.todosLoadErrorMessage = "";
-  hooks.renderTodos?.();
+  EventBus.dispatch("todos:changed", { reason: "todos-loading" });
 
   try {
     const queryParams = buildTodosQueryParams();
@@ -210,7 +211,7 @@ async function loadTodos() {
       state.todosLoadErrorMessage = "";
       state.homeTopFocusState = createInitialHomeTopFocusState();
       await refreshVisibleTodosIfNeeded();
-      hooks.renderTodos?.();
+      EventBus.dispatch("todos:changed", { reason: "todos-loaded" });
       hooks.refreshProjectCatalog?.();
     } else {
       state.todos = [];
@@ -219,7 +220,7 @@ async function loadTodos() {
       clearVisibleTodosState();
       state.todosLoadState = "error";
       state.todosLoadErrorMessage = "Couldn't load tasks";
-      hooks.renderTodos?.();
+      EventBus.dispatch("todos:changed", { reason: "todos-load-error" });
       hooks.refreshProjectCatalog?.();
       hooks.showMessage?.("todosMessage", "Failed to load todos", "error");
     }
@@ -230,7 +231,7 @@ async function loadTodos() {
     clearVisibleTodosState();
     state.todosLoadState = "error";
     state.todosLoadErrorMessage = "Couldn't load tasks";
-    hooks.renderTodos?.();
+    EventBus.dispatch("todos:changed", { reason: "todos-load-error" });
     hooks.refreshProjectCatalog?.();
     console.error("Load todos error:", error);
   }
@@ -285,7 +286,7 @@ async function addTodo() {
       const newTodo = await response.json();
       state.todos.unshift(newTodo);
       await refreshVisibleTodosIfNeeded();
-      hooks.renderTodos?.();
+      EventBus.dispatch("todos:changed", { reason: "todo-added" });
       hooks.updateCategoryFilter?.();
       hooks.clearOnCreateDismissed?.(newTodo.id);
 
@@ -351,7 +352,7 @@ async function toggleTodo(id, forceValue = null) {
       const updatedTodo = await response.json();
       state.todos = state.todos.map((t) => (t.id === id ? updatedTodo : t));
       await refreshVisibleTodosIfNeeded();
-      hooks.renderTodos?.();
+      EventBus.dispatch("todos:changed", { reason: "todo-toggled" });
 
       if (forceValue === null && newCompletedValue) {
         addUndoAction("complete", { id }, "Todo marked as complete");
@@ -378,7 +379,7 @@ async function deleteTodo(id) {
     if (response && response.ok) {
       state.todos = state.todos.filter((t) => t.id !== id);
       state.selectedTodos.delete(id);
-      hooks.renderTodos?.();
+      EventBus.dispatch("todos:changed", { reason: "todo-deleted" });
       hooks.updateCategoryFilter?.();
 
       addUndoAction("delete", todoData, "Todo deleted");
@@ -431,7 +432,7 @@ async function moveTodoToProject(todoId, projectValue) {
     hooks.refreshProjectCatalog?.();
     await hooks.loadProjects?.();
     await refreshVisibleTodosIfNeeded();
-    hooks.renderTodos?.();
+    EventBus.dispatch("todos:changed", { reason: "todo-updated" });
   } catch (error) {
     console.error("Move todo project failed:", error);
     hooks.showMessage?.(
@@ -529,7 +530,7 @@ async function reorderTodos(draggedId, targetId, options = {}) {
     todo.order = index;
   });
 
-  hooks.renderTodos?.();
+  EventBus.dispatch("todos:changed", { reason: "todos-reordered" });
 
   try {
     const response = await hooks.apiCall(`${hooks.API_URL}/todos/reorder`, {
@@ -579,7 +580,7 @@ function toggleSelectAll() {
     filteredTodos.forEach((todo) => state.selectedTodos.delete(todo.id));
   }
 
-  hooks.renderTodos?.();
+  EventBus.dispatch("todos:changed", { reason: "bulk-action" });
 }
 
 function updateSelectAllCheckbox() {
@@ -644,7 +645,7 @@ async function completeSelected() {
 
   state.selectedTodos.clear();
   await refreshVisibleTodosIfNeeded();
-  hooks.renderTodos?.();
+  EventBus.dispatch("todos:changed", { reason: "bulk-action" });
 }
 
 async function deleteSelected() {
@@ -698,7 +699,7 @@ async function deleteSelected() {
   }
 
   state.selectedTodos.clear();
-  hooks.renderTodos?.();
+  EventBus.dispatch("todos:changed", { reason: "bulk-action" });
   hooks.updateCategoryFilter?.();
   if (deletedCount > 0) {
     await loadTodos();
@@ -806,7 +807,7 @@ async function restoreTodo(todoData) {
         return aOrder - bOrder;
       });
       await refreshVisibleTodosIfNeeded();
-      hooks.renderTodos?.();
+      EventBus.dispatch("todos:changed", { reason: "undo-applied" });
       hooks.updateCategoryFilter?.();
     }
   } catch (error) {
